+++ examples/private_service_connect/main.tf
@@ -21,7 +21,9 @@
 # example may fail until that is done.
 
 module "private_service_connect" {
-  source                     = "../../modules/private-service-connect"
+  source  = "terraform-google-modules/network/google//modules/private-service-connect"
+  version = "~> 9.0"
+
   project_id                 = var.project_id
   network_self_link          = module.simple_vpc.network_self_link
   private_service_connect_ip = "10.3.0.5"
@@ -30,7 +32,7 @@ module "private_service_connect" {
 
 module "simple_vpc" {
   source       = "terraform-google-modules/network/google"
-  version      = "~> 7.0"
+  version      = "~> 9.0"
   project_id   = var.project_id
   network_name = "my-custom-network"
   mtu          = 1460