+++ examples/private_service_connect_google_apis/versions.tf
@@ -18,8 +18,8 @@ terraform {
   required_version = ">=0.12.6"
 
   required_providers {
-    google-beta = {
-      version = "~> 4.0"
+    google = {
+      source = "hashicorp/google"
     }
   }
 }