+++ examples/secondary_ranges/main.tf
@@ -22,7 +22,9 @@ locals {
 }
 
 module "vpc-secondary-ranges" {
-  source       = "../../"
+  source  = "terraform-google-modules/network/google"
+  version = "~> 9.0"
+
   project_id   = var.project_id
   network_name = var.network_name