+++ examples/simple_ipv6_project/main.tf
@@ -22,18 +22,20 @@
 
 # [START vpc_custom_create]
 module "test-vpc-module" {
-  source       = "../../"
+  source  = "terraform-google-modules/network/google"
+  version = "~> 9.0"
+
   project_id   = var.project_id # Replace this with your project ID in quotes
   network_name = "my-custom-mode-network"
   mtu          = 1460
 
   subnets = [
     {
-      subnet_name   = "subnet-01"
-      subnet_ip     = "10.10.10.0/24"
-      subnet_region = "us-west1"
-      stack         = "IPV4_IPV6"
-      ipv6_type     = "EXTERNAL"
+      subnet_name      = "subnet-01"
+      subnet_ip        = "10.10.10.0/24"
+      subnet_region    = "us-west1"
+      stack_type       = "IPV4_IPV6"
+      ipv6_access_type = "EXTERNAL"
     },
     {
       subnet_name           = "subnet-02"