+++ examples/simple_project_with_regional_network/main.tf
@@ -20,7 +20,9 @@ locals {
 }
 
 module "test-vpc-module" {
-  source       = "../../"
+  source  = "terraform-google-modules/network/google"
+  version = "~> 9.0"
+
   project_id   = var.project_id
   network_name = var.network_name
   routing_mode = "REGIONAL"