+++ main.tf
@@ -18,15 +18,18 @@
 	VPC configuration
  *****************************************/
 module "vpc" {
-  source                                 = "./modules/vpc"
-  network_name                           = var.network_name
-  auto_create_subnetworks                = var.auto_create_subnetworks
-  routing_mode                           = var.routing_mode
-  project_id                             = var.project_id
-  description                            = var.description
-  shared_vpc_host                        = var.shared_vpc_host
-  delete_default_internet_gateway_routes = var.delete_default_internet_gateway_routes
-  mtu                                    = var.mtu
+  source                                    = "./modules/vpc"
+  network_name                              = var.network_name
+  auto_create_subnetworks                   = var.auto_create_subnetworks
+  routing_mode                              = var.routing_mode
+  project_id                                = var.project_id
+  description                               = var.description
+  shared_vpc_host                           = var.shared_vpc_host
+  delete_default_internet_gateway_routes    = var.delete_default_internet_gateway_routes
+  mtu                                       = var.mtu
+  enable_ipv6_ula                           = var.enable_ipv6_ula
+  internal_ipv6_range                       = var.internal_ipv6_range
+  network_firewall_policy_enforcement_order = var.network_firewall_policy_enforcement_order
 }
 
 /******************************************
@@ -59,6 +62,7 @@ locals {
     for f in var.firewall_rules : {
       name                    = f.name
       direction               = f.direction
+      disabled                = lookup(f, "disabled", null)
       priority                = lookup(f, "priority", null)
       description             = lookup(f, "description", null)
       ranges                  = lookup(f, "ranges", null)
@@ -74,8 +78,10 @@ locals {
 }
 
 module "firewall_rules" {
-  source       = "./modules/firewall-rules"
-  project_id   = var.project_id
-  network_name = module.vpc.network_name
-  rules        = local.rules
+  source        = "./modules/firewall-rules"
+  project_id    = var.project_id
+  network_name  = module.vpc.network_name
+  rules         = local.rules
+  ingress_rules = var.ingress_rules
+  egress_rules  = var.egress_rules
 }