+++ modules/fabric-net-svpc-access/versions.tf
@@ -20,18 +20,18 @@ terraform {
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = "<5.0,>= 2.12"
+      version = ">= 2.12, < 6"
     }
     google-beta = {
       source  = "hashicorp/google-beta"
-      version = "<5.0,>= 2.12"
+      version = ">= 2.12, < 6"
     }
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-network:fabric-net-svpc-access/v7.0.0"
+    module_name = "blueprints/terraform/terraform-google-network:fabric-net-svpc-access/v9.1.0"
   }
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-network:fabric-net-svpc-access/v7.0.0"
+    module_name = "blueprints/terraform/terraform-google-network:fabric-net-svpc-access/v9.1.0"
   }
 }