+++ modules/routes/versions.tf
@@ -20,11 +20,11 @@ terraform {
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = "<5.0,>= 3.83"
+      version = ">= 3.83, < 6"
     }
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-network:routes/v7.0.0"
+    module_name = "blueprints/terraform/terraform-google-network:routes/v9.1.0"
   }
 }