+++ modules/subnets-beta/versions.tf
@@ -20,11 +20,11 @@ terraform {
   required_providers {
     google-beta = {
       source  = "hashicorp/google-beta"
-      version = "<5.0,>= 2.19"
+      version = ">= 2.19, < 6"
     }
   }
 
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-network:subnets-beta/v7.0.0"
+    module_name = "blueprints/terraform/terraform-google-network:subnets-beta/v9.1.0"
   }
 }