+++ variables.tf
@@ -37,7 +37,24 @@ variable "shared_vpc_host" {
 }
 
 variable "subnets" {
-  type        = list(map(string))
+  type = list(object({
+    subnet_name                      = string
+    subnet_ip                        = string
+    subnet_region                    = string
+    subnet_private_access            = optional(string)
+    subnet_private_ipv6_access       = optional(string)
+    subnet_flow_logs                 = optional(string)
+    subnet_flow_logs_interval        = optional(string)
+    subnet_flow_logs_sampling        = optional(string)
+    subnet_flow_logs_metadata        = optional(string)
+    subnet_flow_logs_filter          = optional(string)
+    subnet_flow_logs_metadata_fields = optional(list(string))
+    description                      = optional(string)
+    purpose                          = optional(string)
+    role                             = optional(string)
+    stack_type                       = optional(string)
+    ipv6_access_type                 = optional(string)
+  }))
   description = "The list of subnets being created"
 }
 
@@ -54,8 +71,31 @@ variable "routes" {
 }
 
 variable "firewall_rules" {
-  type        = any
-  description = "List of firewall rules"
+  type = list(object({
+    name                    = string
+    description             = optional(string, null)
+    direction               = optional(string, "INGRESS")
+    disabled                = optional(bool, null)
+    priority                = optional(number, null)
+    ranges                  = optional(list(string), [])
+    source_tags             = optional(list(string))
+    source_service_accounts = optional(list(string))
+    target_tags             = optional(list(string))
+    target_service_accounts = optional(list(string))
+
+    allow = optional(list(object({
+      protocol = string
+      ports    = optional(list(string))
+    })), [])
+    deny = optional(list(object({
+      protocol = string
+      ports    = optional(list(string))
+    })), [])
+    log_config = optional(object({
+      metadata = string
+    }))
+  }))
+  description = "This is DEPRICATED and available for backward compatiblity. Use ingress_rules and egress_rules variables. List of firewall rules"
   default     = []
 }
 
@@ -82,4 +122,80 @@ variable "mtu" {
   type        = number
   description = "The network MTU (If set to 0, meaning MTU is unset - defaults to '1460'). Recommended values: 1460 (default for historic reasons), 1500 (Internet default), or 8896 (for Jumbo packets). Allowed are all values in the range 1300 to 8896, inclusively."
   default     = 0
+}
+
+variable "ingress_rules" {
+  description = "List of ingress rules. This will be ignored if variable 'rules' is non-empty"
+  default     = []
+  type = list(object({
+    name                    = string
+    description             = optional(string, null)
+    disabled                = optional(bool, null)
+    priority                = optional(number, null)
+    destination_ranges      = optional(list(string), [])
+    source_ranges           = optional(list(string), [])
+    source_tags             = optional(list(string))
+    source_service_accounts = optional(list(string))
+    target_tags             = optional(list(string))
+    target_service_accounts = optional(list(string))
+
+    allow = optional(list(object({
+      protocol = string
+      ports    = optional(list(string))
+    })), [])
+    deny = optional(list(object({
+      protocol = string
+      ports    = optional(list(string))
+    })), [])
+    log_config = optional(object({
+      metadata = string
+    }))
+  }))
+}
+
+variable "egress_rules" {
+  description = "List of egress rules. This will be ignored if variable 'rules' is non-empty"
+  default     = []
+  type = list(object({
+    name                    = string
+    description             = optional(string, null)
+    disabled                = optional(bool, null)
+    priority                = optional(number, null)
+    destination_ranges      = optional(list(string), [])
+    source_ranges           = optional(list(string), [])
+    source_tags             = optional(list(string))
+    source_service_accounts = optional(list(string))
+    target_tags             = optional(list(string))
+    target_service_accounts = optional(list(string))
+
+    allow = optional(list(object({
+      protocol = string
+      ports    = optional(list(string))
+    })), [])
+    deny = optional(list(object({
+      protocol = string
+      ports    = optional(list(string))
+    })), [])
+    log_config = optional(object({
+      metadata = string
+    }))
+  }))
+}
+
+variable "enable_ipv6_ula" {
+  type        = bool
+  description = "Enabled IPv6 ULA, this is a permenant change and cannot be undone! (default 'false')"
+  default     = false
+}
+
+variable "internal_ipv6_range" {
+  type        = string
+  default     = null
+  description = "When enabling IPv6 ULA, optionally, specify a /48 from fd20::/20 (default null)"
+}
+
+variable "network_firewall_policy_enforcement_order" {
+  type        = string
+  default     = null
+  description = "Set the order that Firewall Rules and Firewall Policies are evaluated. Valid values are `BEFORE_CLASSIC_FIREWALL` and `AFTER_CLASSIC_FIREWALL`. (default null or equivalent to `AFTER_CLASSIC_FIREWALL`)"
 }