+++ versions.tf
@@ -15,19 +15,19 @@
  */
 
 terraform {
-  required_version = ">= 0.13.0"
+  required_version = ">= 1.3"
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = "< 5.0, >= 3.83"
+      version = ">= 4.64, < 6"
     }
     google-beta = {
       source  = "hashicorp/google-beta"
-      version = "< 5.0, >= 3.45"
+      version = ">= 4.64, < 6"
     }
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-network/v7.0.0"
+    module_name = "blueprints/terraform/terraform-google-network/v9.1.0"
   }
 }